adjust the ratio between angle and stepping of the stepper motor
(cherry picked from commit 42f9032f2796e1a09f4f7e7aca29745f3a4b58cf)

+++ src/vehicle/vehiclecomponent/StepperComt.h
@@ -14,7 +14,7 @@
 class StepperComt : public VehicleComt {
 public:
     StepperComt(uint16_t stepNumber, uint8_t pin1, uint8_t pin2, uint8_t pin3, uint8_t pin4,
-                const String &acIdent, const String &stIdent, const String &rIdent, uint8_t multi,
+                const String &acIdent, const String &stIdent, const String &rIdent, double multi,
                 long speed);
 
 private:
